+++ Editor/Analytics.cs
@@ -17,11 +17,11 @@ namespace UnityEditor.Performance.ProfileAnalyzer
 
         public static void EnableAnalytics()
         {
-            s_EnableAnalytics = true;
-
+#if UNITY_2018_1_OR_NEWER
             AnalyticsResult result = EditorAnalytics.RegisterEventWithLimit(k_EventTopicName, k_MaxEventsPerHour, k_MaxEventItems, k_VendorKey);
-            if (result != AnalyticsResult.Ok)
-                s_EnableAnalytics = false;
+            if (result == AnalyticsResult.Ok)
+                s_EnableAnalytics = true;
+#endif
         }
 
         public enum UIButton
@@ -195,6 +195,7 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             if (!s_EnableAnalytics)
                 return false;
 
+#if UNITY_2018_1_OR_NEWER
             // Duration is in "ticks" 100 nanosecond intervals. I.e. 0.1 microseconds
             float durationInTicks = SecondsToTicks(durationInSeconds);
 
@@ -224,11 +225,15 @@ namespace UnityEditor.Performance.ProfileAnalyzer
                     return false;
             }
 
+
             AnalyticsResult result = EditorAnalytics.SendEventWithLimit(k_EventTopicName, uiButtonEvent);
             if (result != AnalyticsResult.Ok)
                 return false;
 
             return true;
+#else
+            return false;
+#endif
         }
 
         public static bool SendUIUsageModeEvent(UIUsageMode uiUsageMode, float durationInSeconds)
@@ -236,6 +241,7 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             if (!s_EnableAnalytics)
                 return false;
 
+#if UNITY_2018_1_OR_NEWER
             // Duration is in "ticks" 100 nanosecond intervals. I.e. 0.1 microseconds
             float durationInTicks = SecondsToTicks(durationInSeconds);
 
@@ -253,11 +259,15 @@ namespace UnityEditor.Performance.ProfileAnalyzer
                     return false;
             }
 
+
             AnalyticsResult result = EditorAnalytics.SendEventWithLimit(k_EventTopicName, uiUsageEvent);
             if (result != AnalyticsResult.Ok)
                 return false;
 
             return true;
+#else
+            return false;
+#endif
         }
 
         public static bool SendUIVisibilityEvent(UIVisibility uiVisibility, float durationInSeconds, bool show)
@@ -265,6 +275,7 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             if (!s_EnableAnalytics)
                 return false;
 
+#if UNITY_2018_1_OR_NEWER
             // Duration is in "ticks" 100 nanosecond intervals. I.e. 0.1 microseconds
             float durationInTicks = SecondsToTicks(durationInSeconds);
 
@@ -299,6 +310,9 @@ namespace UnityEditor.Performance.ProfileAnalyzer
                 return false;
 
             return true;
+#else
+            return false;
+#endif
         }
 
         public static bool SendUIResizeEvent(UIResizeView uiResizeView, float durationInSeconds, float width, float height, bool isDocked)
@@ -306,6 +320,7 @@ namespace UnityEditor.Performance.ProfileAnalyzer
             if (!s_EnableAnalytics)
                 return false;
 
+#if UNITY_2018_1_OR_NEWER
             // Duration is in "ticks" 100 nanosecond intervals. I.e. 0.1 microseconds
             float durationInTicks = SecondsToTicks(durationInSeconds);
 
@@ -323,11 +338,15 @@ namespace UnityEditor.Performance.ProfileAnalyzer
                     Debug.LogFormat("SendUIResizeEvent: Unsupported view : {0}", uiResizeView);
                     return false;
             }
+
             AnalyticsResult result = EditorAnalytics.SendEventWithLimit(k_EventTopicName, uiResizeEvent);
             if (result != AnalyticsResult.Ok)
                 return false;
 
             return true;
+#else
+            return false;
+#endif
         }