+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0-preview.5] - 2019-04-2
+
+### Updated package.json file to indicate this package is valid for all unity versions
+
+## [0.4.0-preview.4] - 2019-04-2
+
+### Fixed issue in 2017.4 with unsupported analytics API and a GUI style.
+
 ## [0.4.0-preview.3] - 2019-04-1
 
 ### First public release of * Profile Analyzer *.